+++ dflib-avro/src/test/java/org/dflib/avro/AvroLoader_CardinalityTest.java
@@ -23,7 +23,7 @@ public class AvroLoader_CardinalityTest {
     }
 
     @Test
-    public void valueCardinality_Nulls() {
+    public void defaultCardinality_Nulls() {
         DataFrame df = new AvroLoader().load(_TEST1_AVRO);
 
         new DataFrameAsserts(df, "a", "b")
@@ -44,7 +44,7 @@ public class AvroLoader_CardinalityTest {
     }
 
     @Test
-    public void valueCardinality_noNulls() {
+    public void defaultCardinality_noNulls() {
         DataFrame df = new AvroLoader().load(_TEST2_AVRO);
 
         new DataFrameAsserts(df, "a", "b")
@@ -63,10 +63,30 @@ public class AvroLoader_CardinalityTest {
     }
 
     @Test
-    public void valueCardinality_compactCol_Name() {
+    public void compactCols_All() {
+        DataFrame df = new AvroLoader().compactCols().load(_TEST1_AVRO);
+
+        new DataFrameAsserts(df, "a", "b")
+                .expectHeight(6)
+                .expectRow(0, 1, "ab")
+                .expectRow(1, 40000, "ab")
+                .expectRow(2, 40000, "bc")
+                .expectRow(3, 30000, "bc")
+                .expectRow(4, 30000, null)
+                .expectRow(5, null, "bc");
+
+        DataFrame idCardinality = df.cols().select(
+                $col("a").mapVal(System::identityHashCode),
+                $col("b").mapVal(System::identityHashCode));
+
+        assertEquals(4, idCardinality.getColumn(0).unique().size());
+        assertEquals(3, idCardinality.getColumn(1).unique().size());
+    }
+
+    @Test
+    public void compactCols_Name() {
         DataFrame df = new AvroLoader()
-                .compactCol("a")
-                .compactCol("b")
+                .compactCols("a", "b")
                 .load(_TEST1_AVRO);
 
         new DataFrameAsserts(df, "a", "b")
@@ -87,10 +107,9 @@ public class AvroLoader_CardinalityTest {
     }
 
     @Test
-    public void valueCardinality_Nulls_compactCol_Pos() {
+    public void compactCols_Pos_Nulls() {
         DataFrame df = new AvroLoader()
-                .compactCol(0)
-                .compactCol(1)
+                .compactCols(0, 1)
                 .load(_TEST1_AVRO);
 
         new DataFrameAsserts(df, "a", "b")