+++ dflib-parquet/src/test/java/org/dflib/parquet/ParquetLoader_CardinalityTest.java
@@ -122,12 +122,40 @@ public class ParquetLoader_CardinalityTest {
     @Test
     public void compactCardinality() {
         DataFrame df = new ParquetLoader()
-                .compactCol("a")
-                .compactCol("b")
-                .compactCol("c")
-                .compactCol("d")
-                .compactCol("e")
-                .compactCol("f")
+                .compactCols("a", "b", "c", "d", "e", "f")
+                .load(testFile);
+
+        new DataFrameAsserts(df, "a", "b", "c", "d", "e", "f")
+                .expectHeight(6)
+                .expectRow(0, 1, "ab", new BigDecimal("609.10"), true, null, 0L)
+                .expectRow(1, 40000, "ab", new BigDecimal("12.60"), false, 66L, 66L)
+                .expectRow(2, 40000, "bc", new BigDecimal("609.10"), true, 66L, 66L)
+                .expectRow(3, 30000, "bc", new BigDecimal("12.60"), null, 68_000L, 68_000L)
+                .expectRow(4, 30000, null, new BigDecimal("609.10"), true, -66_000L, -66_000L)
+                .expectRow(5, null, "bc", new BigDecimal("609.10"), true, -66_000L, -66_000L);
+
+        DataFrame idCardinality = df.cols().select(
+                $col("a").mapVal(System::identityHashCode),
+                $col("b").mapVal(System::identityHashCode),
+                $col("c").mapVal(System::identityHashCode),
+                $col("d").mapVal(System::identityHashCode),
+                $col("e").mapVal(System::identityHashCode),
+                $col("f").mapVal(System::identityHashCode));
+
+        assertEquals(4, idCardinality.getColumn(0).unique().size());
+        assertEquals(3, idCardinality.getColumn(1).unique().size());
+        assertEquals(2, idCardinality.getColumn(2).unique().size());
+        assertEquals(3, idCardinality.getColumn(3).unique().size());
+        assertEquals(4, idCardinality.getColumn(4).unique().size());
+
+        // primitive columns are resolved without DFLib cache, only using Java Long.valueOf(..) cache for smaller values
+        assertEquals(5, idCardinality.getColumn(5).unique().size());
+    }
+
+    @Test
+    public void compactCardinality_AllCols() {
+        DataFrame df = new ParquetLoader()
+                .compactCols()
                 .load(testFile);
 
         new DataFrameAsserts(df, "a", "b", "c", "d", "e", "f")